Make ‘Stop’ button bring us to a full stop
Originally it was more a pause than a stop. And when paused,
the rest of the interface was essentially dead. The only thing
you could do was press ‘Go’ to resume. Other than reloading the
page, there was no way to abandon a puzzle solution and start over.

With the new behavior the stop button is a full reset, blanking all
fields and requiring new word inputs.